Time: 0 ms (100%), Space: 11.9 MB (7.11%) - LeetHub

+++ 1513-number-of-substrings-with-only-1s/1513-number-of-substrings-with-only-1s.cpp
@@ -8,7 +8,7 @@ public:
                 while(curr < s.size() && s[curr] == '1') {
                     curr++;
                 }
-                int diff = curr - i;
+                long long int diff = curr - i;
                 ans = (ans + (diff * (diff + 1) / 2)) % MOD;
                 i = curr;
             }
